rename servant-jquery to servant-js, Servant.JQuery to Servant.JS
