+++ servant-js/examples/counter.hs
@@ -1,14 +1,3 @@
-# servant-jquery
-
-![servant](https://raw.githubusercontent.com/haskell-servant/servant/master/servant.png)
-
-This library lets you derive automatically (JQuery based) Javascript functions that let you query each endpoint of a *servant* webservice.
-
-## Example
-
-Read more about the following example [here](https://github.com/haskell-servant/servant/tree/master/servant-jquery/tree/master/examples#examples).
-
-``` haskell
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE DeriveGeneric #-}
@@ -21,7 +10,11 @@ import Data.Proxy
 import GHC.Generics
 import Network.Wai.Handler.Warp (run)
 import Servant
-import Servant.JQuery
+import Servant.JS
+import qualified Servant.JS as SJS
+import qualified Servant.JS.Vanilla as JS
+import qualified Servant.JS.JQuery as JQ
+import qualified Servant.JS.Angular as NG
 import System.FilePath
 
 -- * A simple Counter data type
@@ -48,9 +41,9 @@ currentValue :: MonadIO m => TVar Counter -> m Counter
 currentValue counter = liftIO $ readTVarIO counter
 
 -- * Our API type
-type TestApi = "counter" :> Post Counter -- endpoint for increasing the counter
-          :<|> "counter" :> Get  Counter -- endpoint to get the current value
-          :<|> Raw                       -- used for serving static files
+type TestApi = "counter" :> Post '[JSON] Counter -- endpoint for increasing the counter
+          :<|> "counter" :> Get '[JSON] Counter -- endpoint to get the current value
+          :<|> Raw                       -- used for serving static files 
 
 testApi :: Proxy TestApi
 testApi = Proxy
@@ -73,17 +66,31 @@ runServer :: TVar Counter -- ^ shared variable for the counter
 runServer var port = run port (serve testApi $ server var)
 
 -- * Generating the JQuery code
+incCounterJS :: AjaxReq
+currentValueJS :: AjaxReq
+incCounterJS :<|> currentValueJS :<|> _ = javascript testApi
 
-incCounterJS :<|> currentValueJS :<|> _ = jquery testApi
-
-writeJS :: FilePath -> [AjaxReq] -> IO ()
-writeJS fp functions = writeFile fp $
-  concatMap generateJS functions
-
+writeJS :: JavaScriptGenerator -> FilePath -> [AjaxReq] -> IO ()
+writeJS gen fp functions = writeFile fp $
+  concatMap (\req -> generateJS req gen) functions
+  
+writeServiceJS :: FilePath -> [AjaxReq] -> IO ()
+writeServiceJS fp functions = writeFile fp $
+  NG.wrapInServiceWith (NG.defAngularOptions { NG.serviceName = "counterSvc" })
+    (defCommonGeneratorOptions { SJS.moduleName = "counterApp" }) functions
+  
 main :: IO ()
 main = do
   -- write the JS code to www/api.js at startup
-  writeJS (www </> "api.js")
+  writeJS JQ.generateJQueryJS (www </> "jquery" </> "api.js")
+          [ incCounterJS, currentValueJS ]
+  writeJS JS.generateVanillaJS (www </> "vanilla" </> "api.js")
+          [ incCounterJS, currentValueJS ]
+  writeJS (NG.generateAngularJS 
+            NG.defAngularOptions) (www </> "angular" </> "api.js")
+          [ incCounterJS, currentValueJS ]
+  writeServiceJS 
+          (www </> "angular" </> "api.service.js")
           [ incCounterJS, currentValueJS ]
 
   -- setup a shared counter
@@ -91,4 +98,3 @@ main = do
 
   -- listen to requests on port 8080
   runServer cnt 8080
-```
