+++ servant-js/examples/www/angular/index.html
@@ -7,25 +7,25 @@
   #inc { margin: 0px 20px; background-color: green; color: white; }
   </style>
   <script src="angular.min.js" type="text/javascript"></script>
+  <script src="api.js" type="text/javascript"></script>
 </head>
 <body ng-controller="CounterCtrl">
-<h1>Angular version (Service version)</h1>
+<h1>Angular version</h1>
 <span id="counter">{{ 'Counter: ' + counter }}</span>
 <button ng-click="incCounter()" id="inc">Increase</button>
-or <a href="/doc">view the docs</a>
 <script type="text/javascript">
 var counterApp = angular.module('counterApp', []);
 
-counterApp.controller('CounterCtrl', function ($scope,counterSvc, $interval) {
+counterApp.controller('CounterCtrl', ['$scope', '$http', '$interval', function ($scope,$http, $interval) {
   // Let's define how we publish information to the $scope
   var publishCounter = function (response) { $scope.counter = response.value; };
   
   $scope.getCounter = function() {
-    counterSvc.getcounter()
+    getcounter($http)
         .success(publishCounter);
   }
   $scope.incCounter = function() {
-    counterSvc.postcounter()
+    postcounter($http)
         .success(publishCounter);
   }
   
@@ -33,8 +33,7 @@ counterApp.controller('CounterCtrl', function ($scope,counterSvc, $interval) {
   $scope.getCounter();
   // we update the value every 1sec, in the same way
   $interval($scope.getCounter, 1000);
-});
+}]);
 </script>
-<script src="api.service.js" type="text/javascript"></script>
 </body>
 </html>
