+++ servant-js/examples/www/angular/service.html
@@ -7,26 +7,24 @@
   #inc { margin: 0px 20px; background-color: green; color: white; }
   </style>
   <script src="angular.min.js" type="text/javascript"></script>
-  <script src="api.js" type="text/javascript"></script>
 </head>
 <body ng-controller="CounterCtrl">
-<h1>Angular version</h1>
+<h1>Angular version (Service version)</h1>
 <span id="counter">{{ 'Counter: ' + counter }}</span>
 <button ng-click="incCounter()" id="inc">Increase</button>
-or <a href="/doc">view the docs</a>
 <script type="text/javascript">
 var counterApp = angular.module('counterApp', []);
 
-counterApp.controller('CounterCtrl', ['$scope', '$http', '$interval', function ($scope,$http, $interval) {
+counterApp.controller('CounterCtrl', function ($scope,counterSvc, $interval) {
   // Let's define how we publish information to the $scope
   var publishCounter = function (response) { $scope.counter = response.value; };
   
   $scope.getCounter = function() {
-    getcounter($http)
+    counterSvc.getcounter()
         .success(publishCounter);
   }
   $scope.incCounter = function() {
-    postcounter($http)
+    counterSvc.postcounter()
         .success(publishCounter);
   }
   
@@ -34,7 +32,8 @@ counterApp.controller('CounterCtrl', ['$scope', '$http', '$interval', function (
   $scope.getCounter();
   // we update the value every 1sec, in the same way
   $interval($scope.getCounter, 1000);
-}]);
+});
 </script>
+<script src="api.service.js" type="text/javascript"></script>
 </body>
 </html>
