+++ servant-js/examples/www/jquery/index.html
@@ -8,14 +8,14 @@
   </style>
 </head>
 <body>
-<h1>Vanilla version</h1>
+<h1>JQuery version</h1>
 <span id="counter">Counter: 0</span>
 <button id="inc">Increase</button>
-or <a href="/doc">view the docs</a>
 
+<script src="jquery.min.js" type="text/javascript"></script>
 <script src="api.js" type="text/javascript"></script>
 <script type="text/javascript">
-window.addEventListener('load', function() {
+$(document).ready(function() {
   // we get the current value stored by the server when the page is loaded
   getcounter(updateCounter, alert);
 
@@ -27,12 +27,12 @@ window.addEventListener('load', function() {
 
 function updateCounter(response)
 {
-  document.getElementById('counter').innerHTML = 'Counter: ' + response.value;
+  $('#counter').html('Counter: ' + response.value);
 }
 
 // when the button is clicked, ask the server to increase
 // the value by one
-document.getElementById('inc').addEventListener('click', function() {
+$('#inc').click(function() {
   postcounter(updateCounter, alert);
 });
 </script>
