+++ servant-js/examples/www/vanilla/index.html
@@ -8,15 +8,13 @@
   </style>
 </head>
 <body>
-<h1>JQuery version</h1>
+<h1>Vanilla version</h1>
 <span id="counter">Counter: 0</span>
 <button id="inc">Increase</button>
-or <a href="/doc">view the docs</a>
 
-<script src="jquery.min.js" type="text/javascript"></script>
 <script src="api.js" type="text/javascript"></script>
 <script type="text/javascript">
-$(document).ready(function() {
+window.addEventListener('load', function() {
   // we get the current value stored by the server when the page is loaded
   getcounter(updateCounter, alert);
 
@@ -28,12 +26,12 @@ $(document).ready(function() {
 
 function updateCounter(response)
 {
-  $('#counter').html('Counter: ' + response.value);
+  document.getElementById('counter').innerHTML = 'Counter: ' + response.value;
 }
 
 // when the button is clicked, ask the server to increase
 // the value by one
-$('#inc').click(function() {
+document.getElementById('inc').addEventListener('click', function() {
   postcounter(updateCounter, alert);
 });
 </script>
