+++ servant-js/src/Servant/JS/Vanilla.hs
@@ -1,6 +1,6 @@
-module Servant.JQuery.Vanilla where
+module Servant.JS.Vanilla where
 
-import Servant.JQuery.Internal
+import Servant.JS.Internal
 import Control.Lens
 import Data.List
 import Data.Monoid
