+++ servant-js/test/Servant/JSSpec.hs
@@ -5,7 +5,7 @@
 {-# LANGUAGE TypeFamilies        #-}
 {-# LANGUAGE TypeOperators       #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
-module Servant.JQuerySpec where
+module Servant.JSSpec where
 
 import Data.Either (isRight)
 import Data.Proxy
@@ -13,11 +13,11 @@ import Language.ECMAScript3.Parser (parseFromString)
 import Test.Hspec
 
 import Servant.API
-import Servant.JQuery
-import qualified Servant.JQuery.Vanilla as JS
-import qualified Servant.JQuery.JQuery as JQ
-import qualified Servant.JQuery.Angular as NG
-import Servant.JQuerySpec.CustomHeaders
+import Servant.JS
+import qualified Servant.JS.Vanilla as JS
+import qualified Servant.JS.JQuery as JQ
+import qualified Servant.JS.Angular as NG
+import Servant.JSSpec.CustomHeaders
 
 type TestAPI = "simple" :> ReqBody '[JSON,FormUrlEncoded] String :> Post '[JSON] Bool
           :<|> "has.extension" :> Get '[FormUrlEncoded,JSON] Bool
